requirements.txt
akkaserverless @ file:///Users/jeremypollock/development/lightbend/product/akkaserverless-python-sdk
appier==1.24.0
attrs==21.2.0
bleach==4.0.0
certifi==2021.5.30
charset-normalizer==2.0.4
cloudstate @ file:///Users/jeremypollock/development/lightbend/product/akkaserverless-python-sdk
colorama==0.4.4
deprecation==2.1.0
docker==5.0.0
docutils==0.17.1
google-api==0.1.12
googleapis-common-protos==1.53.0
grpcio==1.39.0
grpcio-reflection==1.39.0
grpcio-tools==1.39.0
idna==3.2
importlib-metadata==4.6.3
iniconfig==1.1.1
keyring==23.0.1
packaging==21.0
pkginfo==1.7.1
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
Pygments==2.9.0
pyparsing==2.4.7
pytest==6.2.4
readme-renderer==29.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
six==1.16.0
testcontainers==3.4.1
toml==0.10.2
tqdm==4.62.0
twine==3.4.2
urllib3==1.26.6
webencodings==0.5.1
websocket-client==1.1.1
wrapt==1.12.1
zipp==3.5.0
